Autoload all the things
This will help prepare to move some of this code into a separate
composer module. Which in turn will allow us to write tests for all of
this code.
