+++ app/Lib/RootsWalkerComment.php
@@ -1,10 +1,14 @@
 <?php
+
+namespace MyTheme\Lib;
+
 /**
  * Use Bootstrap's media object for listing comments
  *
  * @link http://getbootstrap.com/components/#media
  */
-class Roots_Walker_Comment extends Walker_Comment {
+
+class RootsWalkerComment extends \Walker_Comment {
   function start_lvl(&$output, $depth = 0, $args = array()) {
     $GLOBALS['comment_depth'] = $depth + 1; ?>
     <ul <?php comment_class('media list-unstyled comment-' . get_comment_ID()); ?>>
@@ -41,12 +45,17 @@ class Roots_Walker_Comment extends Walker_Comment {
     // Close ".media-body" <div> located in templates/comment.php, and then the comment's <li>
     echo "</div></li>\n";
   }
-}
 
-function roots_get_avatar($avatar, $type) {
-  if (!is_object($type)) { return $avatar; }
+    function roots_get_avatar($avatar, $type) {
+      if (!is_object($type)) { return $avatar; }
 
-  $avatar = str_replace("class='avatar", "class='avatar pull-left media-object", $avatar);
-  return $avatar;
+      $avatar = str_replace("class='avatar", "class='avatar pull-left media-object", $avatar);
+      return $avatar;
+    }
+
+
+    public static function register()
+    {
+        add_filter('get_avatar', ['\\MyTheme\\Lib\\RootsWalkerComment', 'roots_get_avatar'], 10, 2);
+    }
 }
-add_filter('get_avatar', 'roots_get_avatar', 10, 2);
