+++ app/Posts/CustomFields.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace MyTheme\Posts;
+
+class CustomFields
+{
+    public static function register()
+    {
+        // ...
+    }
+}
