+++ app/Posts/PostTypes.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace MyTheme\Posts;
+
+class PostTypes
+{
+    public static function register()
+    {
+        // ...
+    }
+}
