+++ app/Theme/Media.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace MyTheme\Theme;
+
+class Media
+{
+    public static function register()
+    {
+        set_post_thumbnail_size( 150, 150, true );
+        add_image_size( 'medium', 200, 200, true );
+        add_image_size( 'large', 800, 300, true );
+    }
+}
