+++ app/Theme/Menus.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace MyTheme\Theme;
+
+class Menus
+{
+    public static function register()
+    {
+        register_nav_menu( 'header', 'Header Menu' );
+        register_nav_menu( 'footer', 'Footer Menu' );
+    }
+}
