--- app/lib/whippet/whippet.php
@@ -1,9 +1,0 @@
-<?php
-
-//
-//
-//
-
-require __DIR__ . '/layout.php';
-require __DIR__ . '/template_tags.php';
-
