--- app/posts/post_types.php
@@ -1,6 +1,0 @@
-<?php
-
-if(function_exists("register_field_group"))
-{
-
-}
