--- app/theme/media.php
@@ -1,5 +1,0 @@
-<?php
-
-set_post_thumbnail_size( 150, 150, true );
-add_image_size( 'medium', 200, 200, true );
-add_image_size( 'large', 800, 300, true );
