--- app/theme/menus.php
@@ -1,4 +1,0 @@
-<?php
-
-register_nav_menu( 'header', 'Header Menu' );
-register_nav_menu( 'footer', 'Footer Menu' );
