+++ templates/partials/pager.php
@@ -1,7 +1,7 @@
 <?php if ($wp_query->max_num_pages > 1) : ?>
 
   <div class="pager">
-    <?php pagination(); ?>
+    <?php \MyTheme\Theme\Pagination::pagination(); ?>
   </div>
 
 <?php endif; ?>
